Check for `live-blog` content type
If a live blog has a content type of `live-blog` and `realtime` is true,
we know it's a standalone live blog. We don't need to check for the
`liveBlog` property. The `liveBlog` property doesn't come from
Elasticsearch and is very `next-article` specific. This component is
used by both the `ft-app` and `next-article`.